feat: allow legacy behaviour with to prop
