+++ packages/ra-ui-materialui/src/button/Button.tsx
@@ -7,10 +7,11 @@ import {
     useMediaQuery,
     Theme,
 } from '@mui/material';
-
 import { styled } from '@mui/material/styles';
+import { To } from 'history';
 import PropTypes from 'prop-types';
 import { useTranslate } from 'ra-core';
+import { Path } from 'react-router';
 
 /**
  * A generic Button with side icon. Only the icon is displayed on small screens.
@@ -36,10 +37,12 @@ export const Button = <RootComponent extends React.ElementType = 'button'>(
         label,
         color = 'primary',
         size = 'small',
+        to: locationDescriptor,
         ...rest
     } = props;
     const translate = useTranslate();
     const translatedLabel = label ? translate(label, { _: label }) : undefined;
+    const linkParams = getLinkParams(locationDescriptor);
 
     const isXSmall = useMediaQuery((theme: Theme) =>
         theme.breakpoints.down('sm')
@@ -53,6 +56,7 @@ export const Button = <RootComponent extends React.ElementType = 'button'>(
                     className={className}
                     color={color}
                     size="large"
+                    {...linkParams}
                     {...rest}
                 >
                     {children}
@@ -64,6 +68,7 @@ export const Button = <RootComponent extends React.ElementType = 'button'>(
                 color={color}
                 disabled={disabled}
                 size="large"
+                {...linkParams}
                 {...rest}
             >
                 {children}
@@ -78,6 +83,7 @@ export const Button = <RootComponent extends React.ElementType = 'button'>(
             disabled={disabled}
             startIcon={alignIcon === 'left' && children ? children : undefined}
             endIcon={alignIcon === 'right' && children ? children : undefined}
+            {...linkParams}
             {...rest}
         >
             {translatedLabel}
@@ -90,6 +96,7 @@ interface Props<RootComponent extends React.ElementType> {
     children?: React.ReactElement;
     className?: string;
     component?: RootComponent;
+    to?: LocationDescriptor | To;
     disabled?: boolean;
     label?: string;
     size?: 'small' | 'medium' | 'large';
@@ -133,3 +140,30 @@ const StyledButton = styled(MuiButton, {
         lineHeight: 1.5,
     },
 });
+
+export const getLinkParams = (
+    locationDescriptor?: LocationDescriptor | string
+) => {
+    // eslint-disable-next-line eqeqeq
+    if (locationDescriptor == undefined) {
+        return undefined;
+    }
+
+    if (typeof locationDescriptor === 'string') {
+        return { to: locationDescriptor };
+    }
+
+    const { redirect, replace, state, ...to } = locationDescriptor;
+    return {
+        to,
+        redirect,
+        replace,
+        state,
+    };
+};
+
+export type LocationDescriptor = Partial<Path> & {
+    redirect?: boolean;
+    state?: any;
+    replace?: boolean;
+};
